refactor(analytics): factory for internal GA4/GSC report routes (audit consistency)
10 admin-only internal-analytics route handlers (7 GA4 + 3 GSC) were
byte-identical except for the report key and one optional query param
(limit / compare / neither). Add lib/internal-analytics/make-report-route.ts
makeReportRoute(report, extra) — shared admin gate, `range` parsing, and
TaxonomyAwareError handling — and reduce each route to a one-line export.
Report keys unchanged; behavior identical. -159 lines.

Verified: tsc clean (confirms every key is a valid ReportKey), eslint clean,
295 tests pass.
