+++ frontend/app/api/internal/analytics/ga4/upgrade-views/route.ts
@@ -1,22 +1,6 @@
-import { NextResponse } from 'next/server';
-import { requireAdmin } from '@/lib/supabase/admin';
-import { runReport } from '@/lib/internal-analytics/report-registry';
-import { TaxonomyAwareError } from '@/lib/internal-analytics/queries/_error';
+import { makeReportRoute } from '@/lib/internal-analytics/make-report-route';
 
 export const runtime = 'nodejs';
 export const dynamic = 'force-dynamic';
 
-export async function GET(req: Request) {
-  const auth = await requireAdmin();
-  if (auth.error) return auth.error;
-  const url = new URL(req.url);
-  const range = url.searchParams.get('range') ?? 'last_7_days';
-  const compare = url.searchParams.get('compare') === '1';
-  try {
-    const data = await runReport('ga4_upgrade_views', { date_range: range, compare_to_previous: compare });
-    return NextResponse.json(data);
-  } catch (e) {
-    if (e instanceof TaxonomyAwareError) return NextResponse.json({ error: e.taxonomy }, { status: 500 });
-    throw e;
-  }
-}
+export const GET = makeReportRoute('ga4_upgrade_views', 'compare');
